Supports PHP 8.2 and updates dependencies (#2)
* Update README.md
* Update LICENSE
* Adds PHP 8.2 support

Updates the Docker image to PHP 8.2 to ensure compatibility and take advantage of new features and performance improvements.

Also, adds php-codesniffer to the dev requirements to ensure code quality and consistency.

* Supports PHP 8.2 and updates dependencies

Updates the project to support PHP 8.2 and later versions.

This change involves:
- Updating the required PHP version in composer.json.
- Updating the Github Actions workflow to use PHP 8.2, 8.3 and 8.4.
- Updating dependencies in composer.json and creating composer.lock file.
- Fixing minor issues with the test suite.

* Updates CI workflow

Updates the CI workflow to use actions/cache@v4 and adds the curl extension to the PHP environment.

These updates ensure dependencies are cached correctly, and adds the curl extension as a dependency for PHP 8.2 support.

* Updates composer cache directory setting

Updates the way the composer cache directory is set in the CI workflow to be compatible with newer versions of actions/cache.

* Fixes composer cache directory output format

Updates the way the composer cache directory is echoed to
ensure proper parsing by GitHub Actions. This resolves
an issue where the cache directory was not being correctly
identified, potentially leading to unnecessary dependency
downloads.

* Improves composer cache key generation.

Updates the composer cache key to use `composer.lock` instead of `composer.json`.
This ensures the cache is invalidated when dependencies are updated, leading to more accurate and reliable dependency management.